requirements.txt
einops==0.8.0
matplotlib==3.9.2
numpy==2.1.0
Pillow==10.4.0
pycocotools==2.0.8
scipy==1.14.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4